ping: Allow to ping IPv6 link-local scope address without link specification
link-local address required to specify interface either with -I option
option) or '%'-notation in destination.
If not it failed with "connect: Invalid argument".
Now it's possible to ping without it.

NOTE: it used to be possible in the past: added in f68eec0 ("ping:
perform dual-stack ping by default") and removed again in
e25568f ("correctly initialize first hop").

+ allow this for link-local scope multicast address (TODO: is it
convinient?)

Fixes #99.