ping: allocate memory to buffers in run time structure
The cmsgbuf is 4096 bytes of stack, that is wasted when pinging IPv4
addresses.  The rcvd_tbl is always needed, but 8192 is enough stack size
that it is better to use proper allocation function so that possible
allocation failure can be detected.

Secondly making the whole of run time struct to fit into a single memory
page should maximize probability of it to be in cache.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>