ping: Add option -e to set Identifier field for ICMP ECHO_REQUEST
There are limitations, which requires using SOCK_RAW:

* Identifier field can be set on IPv6 only on SOCK_RAW.

* On IPv4 also SOCK_DGRAM can be used unless the value is 0.
To set it 0, SOCK_RAW needs to be used.

Code path in kernel on IPv4:

When using SOCK_DGRAM the identification field is set in when binding to
an address with port set to the value of desired identifier field.
net/ipv4/af_inet.c:__inet_bind() function at some point calls
net/ipv4/ping.c:ping_get_port() (via struct proto ping_prot.get_port)
where value of the port is assigned to isk->inet_num which is later used
fror identifier.

When passing 0 to ping_get_port() it goes to "default path", which is
incrementing global value of identifier field by one and using it.
That's the reason why SOCK_RAW needs to be used.

Closes: https://github.com/iputils/iputils/pull/426

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Tested-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Vadim Rumyantsev <vadelve@yandex.ru>
[ pvorel: fixes: raw socket default for IPv6, man, help, commit message ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>