ping: Fix EMSGSIZE on -s > 65527 on ICMP datagram socket
ICMP datagram socket fails on -s > 65527 on both IPv4 and IPv6,
raw socket fails only on IPv4. This is a kernel limitation ping should
encounter. It's actually a local error, but reusing rts->nerrors to
avoid keeping recvmsg() until timeout (avoid slowdown) and get the error
reported. Maybe it'd be better to stop after first error.

This will be workarounded with later lower down -s max size in next
commits, but better to fix before it in case it gets max size reenabled
in the future.

Signed-off-by: Petr Vorel <pvorel@suse.cz>