ping: Print data bytes
data bytes is in original ping, therefore it should be in JSON as well.

TODO: IPv4 has more info, but that requires ping_func_set_st *fset
pointer - obviously gather_statistics() is not a good place to print the
"header" data. Maybe move it to ping_print_packet(), which is used only
for non-json code (which is sign of wrong design)?

TODO: what would be proper names these "56 (84) data bytes"?

Signed-off-by: Petr Vorel <pvorel@suse.cz>